requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-socio-grpc==0.24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
idna==3.10
inflection==0.5.1
lark==1.2.2
packaging==24.2
protobuf==5.29.3
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
